README.mdの不要な行を削除し、package-lock.jsonに新しい依存関係を追加しました。また、useCovidDataフックを作成し、usePrefectureDataPreparationフックを更新してデータ取得ロジックを整理しました。
